+++ app/Http/Controllers/Reservation/ReservationController.php
@@ -8,7 +8,6 @@ use App\Http\Requests\Reservation\UpdateReservationRequest;
 use App\Http\Resources\Reservation\ReservationResource;
 use App\Models\Reservation\Reservation;
 use App\Services\Reservation\ReservationService;
-use Carbon\Carbon;
 use Illuminate\Http\JsonResponse;
 use Illuminate\Http\Request;
 use Illuminate\Http\Resources\Json\AnonymousResourceCollection;
@@ -22,32 +21,42 @@ class ReservationController extends Controller
 
     public function index(Request $request): AnonymousResourceCollection
     {
+        $this->authorize('viewAny', Reservation::class);
+
         $reservations = $this->reservation->getReservations($request);
 
         return ReservationResource::collection($reservations);
     }
 
-    public function show(Reservation $reservation): JsonResponse
+    public function show(Reservation $reservation): ReservationResource
     {
-        return response()->json($reservation->load('location', 'user'));
+        $this->authorize('view', $reservation);
+
+        return new ReservationResource($reservation->load('location', 'user'));
     }
 
-    public function store(ReservationRequest $request): JsonResponse
+    public function store(ReservationRequest $request): ReservationResource
     {
-        $reservation = $request->user()->reservations()->create($request->all());
+        $this->authorize('create', Reservation::class);
 
-        return response()->json($reservation, 201);
+        $reservation = Reservation::create($request->all());
+
+        return new ReservationResource($reservation->load('location', 'user'));
     }
 
-    public function update(UpdateReservationRequest $request, Reservation $reservation): JsonResponse
+    public function update(UpdateReservationRequest $request, Reservation $reservation): ReservationResource
     {
+        $this->authorize('update', $reservation);
+
         $reservation->update($request->all());
 
-        return response()->json($reservation);
+        return new ReservationResource($reservation);
     }
 
     public function destroy(Reservation $reservation): JsonResponse
     {
+        $this->authorize('delete', $reservation);
+
         $reservation->delete();
 
         return response()->json(['message' => 'Reservation deleted successfully']);
